+++ smach_viewer/setup.py
@@ -4,7 +4,7 @@ from distutils.core import setup
 from catkin_pkg.python_setup import generate_distutils_setup
 
 d = generate_distutils_setup(
-   packages=['smach_viewer_module', 'smach_viewer_module.xdot'],
+   packages=['smach_viewer_module','smach_viewer_module.wxxdot'],
    package_dir={'': 'src'}
 )
