+++ smach_viewer/src/smach_viewer_module/__init__.py
@@ -1,0 +1,3 @@
+# from smach_viewer_module import wxxdot, xdot_qt
+# from wxxdot import *
+from smach_viewer_module.xdot_qt import *
