--- smach_viewer/src/smach_viewer_module/xdot/__init__.py
@@ -1,1 +1,0 @@
-from smach_viewer_module.xdot import wxxdot, xdot_qt
